refactor: share no-mocks-import rule
