+++ internal/plugins/rstest/rules/no_mocks_import/no_mocks_import.go
@@ -1,5 +1,13 @@
 package no_mocks_import
 
-import jestNoMocksImport "github.com/web-infra-dev/rslint/internal/plugins/jest/rules/no_mocks_import"
+import (
+	"github.com/web-infra-dev/rslint/internal/rule"
+	sharedNoMocksImport "github.com/web-infra-dev/rslint/internal/utils/test_framework/rules/no_mocks_import"
+)
 
-var NoMocksImportRule = jestNoMocksImport.NewRule("rstest/no-mocks-import", "rs.mock")
+// NewRule creates a no-mocks-import rule for a test framework.
+func NewRule(name string, mockFunction string) rule.Rule {
+	return sharedNoMocksImport.NewRule(name, mockFunction)
+}
+
+var NoMocksImportRule = NewRule("rstest/no-mocks-import", "rs.mock")
